Switch to using EMBED tag
According to sources, it seems that for ExternalInterface to work with OBJECT,
you'd need to set an attribute for
classid="clsid:D27CDB6E-AE6D-11cf-96B8-444553540000", which invokes the
ActiveX control to load the object. Obviously this will fail on Chrome.
https://stackoverflow.com/a/22354224

The main tfm page already relies on the fact that OBJECT will fail on non-IE
browsers and fallback to using EMBED, no harm using it here.